host: phc: Add host PHC driver for 9xxx
The Octeon host PHC driver uses the Octeon EP PTP clock directly to allow
the host clock to be disciplined by the Octeon PTP clock.  This allows the
host to be synced to the same master clock as the Octeon EP is synced with
over the network.
An additional SDP PCIe PF is created by firmware with a specific PCIe ID
for this.
This driver relies on some common code and headers use by the SDP network
driver, and will need to be separated as part of upstreaming.

IPBUSW-19810
Change-Id: I103ae24a3b07c821eae3198538fcf53e76ae3c65
Signed-off-by: Roy Franz <rfranz@marvell.com>