host: Allow PHC driver to load with FW status of RUNNING (2)
Allow the PHC driver to load with FW_STATUS values of
READY (1) and RUNNING (2).  This allows the PHC driver
to be loaded before or after the SDP network driver.
This driver is independent of the SDP network driver,
and can be loaded in either state.
The PHC driver does not modify the FW_STATUS.

IPBUSW-25269
Signed-off-by: Roy Franz <rfranz@marvell.com>
Change-Id: Ibfe6f5638dee5f4fa6b4d2b105090473c0d84acd