octeon_ep: Fix pci_enable_ptm for RHEL 8.4 compatibility
The pci_enable_ptm is put as private in RHEL 8.4 but it is still
exported. So this patch defines a prototype of that function so that the
module can be compiled and the function can be used.

Signed-off-by: Radha Mohan Chintakuntla <radhac@marvell.com>
Change-Id: Icd9bfd581a4e6fa6ad46cad9395ae02832eff9ff
(cherry picked from commit f33b1e0eb7600ab988b1761b54db7986523b11fd)