octboot_net: mark state transitions more accurately
mark the device as available when the bar4 is accessible again.

also clear signature_found when device becomes unavailable, as it is
implicit signature cannot be found when device is not accessible.
otherwise, signature_found is still read as 'true' while device is
offline and will not be able to identify the transition when device is
back online and will not save pci config state, which results in failure
to restore state when octeon is cold reset again.

Signed-off-by: Veerasenareddy Burru <vburru@marvell.com>
Change-Id: I54e3f65f4087873d1d503dac59227d9cc9d4134f
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/pcie_ep/pcie_ep_octeon_host/+/104150
Tested-by: sa_ip-sw-jenkins <sa_ip-sw-jenkins@marvell.com>
Reviewed-by: Sathesh B Edara <sedara@marvell.com>