octeon_ep: get firmware ready should explicitly test for ready value
The firmware ready value is 1, and get firmware ready should explicitly
test for that value instead of giving true, when read status is greater
than 0. The firmware ready value read will be 2 after driver load, and
on unbind till fw rewrites the firmware ready back to 0, the value
seen by driver will be 2, which should be regarded as not ready.

Signed-off-by: Shinas Rasheed <srasheed@marvell.com>
Change-Id: I6b543f83ab91442475d36446d17284a28b9cd62a
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/pcie_ep/pcie_ep_octeon_host/+/105596
Reviewed-by: Sathesh B Edara <sedara@marvell.com>
Tested-by: sa_ip-sw-jenkins <sa_ip-sw-jenkins@marvell.com>